% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coloring.R
\name{get.input.set.color.vector}
\alias{get.input.set.color.vector}
\title{Find the ISCV (Input Set Color Vector) of all nodes in the graph given the coloring.}
\usage{
get.input.set.color.vector(
  nodes,
  raw_edges = NA,
  file = NA,
  sep = " ",
  header = F,
  directed = F
)
}
\arguments{
\item{nodes}{Variable is formatted as the output of the get.balanced.coloring.Kamei() function. Dataframe with 2 columns: Name, Color.}

\item{raw_edges}{2 or 3 column data frame specifying the list of edges (specify only one file or raw_edges)}

\item{file}{Path to the file with the edgelist. Make sure to specify "sep" (if different from " ") and "header" (if different from FALSE) to be passed to the read.table function.}

\item{sep}{To be used with the "file" variable. Defines the field separator character to be used in the read.table() function. Is set with " " by default.}

\item{header}{To be used with the "file" variable. A logical value indicating whether "file" contains the names of the variables as its first line. Is set as FALSE by default.}

\item{directed}{A logical value indicating whether the network is directed or not. Is set as FALSE by default.}
}
\value{

}
\description{
The get.input.set.color.vector() function finds ISCVs (Input Set Color Vectors as defined in SI VI of (1)) of all nodes in the graph given the coloring of the graph.
Coloring is specified by "nodes" variable formatted as the output of the get.balanced.coloring.Kamei() function.
Input graph is specified using "raw_edges", "file", "sep" and "header" variables.
(1) Leifer I, Morone F, Reis SDS, Andrade JS, Sigman M, Makse HA. Circuits with broken fibration symmetries perform core logic computations in biological networks. PLoS Comput Biol 2020;16(6):e1007776.
}
