% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coloring.R
\name{get.balanced.coloring.Kamei}
\alias{get.balanced.coloring.Kamei}
\title{Find balanced coloring using Kamei algorithm}
\usage{
get.balanced.coloring.Kamei(
  raw_edges = NA,
  file = NA,
  sep = " ",
  header = F,
  directed = F,
  look.for.no.input.nodes = T
)
}
\arguments{
\item{raw_edges}{List of edges}

\item{file}{File from which edgelist can be read (specify only one file or raw_edges)}

\item{sep}{Field separator character used in read.table. If missing is set at " ".}

\item{header}{A logical value indicating whether "file" contains the names of the variables as its first line. If missing is set at F}

\item{directed}{A logical value indicating whether the network is directed or not. If missing is set at F}

\item{look.for.no.input.nodes}{A logical value indicating whether colors of nodes with no inputs need to be fixed. If missing is set at T}
}
\value{
Data frame with 2 columns: Name (for the node name) and Color (for the node color id)
}
\description{
This function finds a balanced coloring. Specify only one file or raw_edges.
}
