% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get.balanced.coloring.Kamei}
\alias{get.balanced.coloring.Kamei}
\title{Find balanced coloring using Kamei algorithm}
\usage{
get.balanced.coloring.Kamei(
  raw_edges = NA,
  file = NA,
  sep = "\\t",
  header = F,
  directed = F,
  weighted = F,
  look.for.no.input.nodes = T
)
}
\arguments{
\item{raw_edges}{list of edges}
}
\value{
Data frame with 2 columns: Name (for the node name) and Color (for )
}
\description{
This function finds a balanced coloring
}
