% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvalues.R
\name{get.building.block.pvalues}
\alias{get.building.block.pvalues}
\title{Find fibration building blocks and their p-values}
\usage{
get.building.block.pvalues(
  raw_edges = NA,
  file = NA,
  sep = " ",
  header = F,
  sampleSize = 10000,
  mode = c("nl", "Class", "BlockName"),
  method = c("degreeSequence", "erdosrenyi")
)
}
\arguments{
\item{raw_edges}{List of edges}

\item{file}{File from which edgelist can be read (specify only one file or raw_edges)}

\item{sep}{Field separator character used in read.table. If missing is set at " ".}

\item{header}{A logical value indicating whether "file" contains the names of the variables as its first line. If missing is set at F.}

\item{sampleSize}{Size of the sample on which to find p-values. Default is 10000.}

\item{mode}{Which p-values need to be found: nl, class or block name.}

\item{method}{Randomization method: "degreeSequence" or "erdosrenyi".}
}
\value{
The count of building blocks in the network, their occurence in the random network and corresponding Z-Scores and p-values.
}
\description{
This function finds fibration building blocks Specify only one file or raw_edges.
}
