% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildingBlocks.R
\name{get.building.blocks}
\alias{get.building.blocks}
\title{Find fibration building blocks}
\usage{
get.building.blocks(
  raw_edges = NA,
  file = NA,
  sep = " ",
  header = F,
  outputFolder = NA,
  csv = F,
  png = F,
  pdf = F,
  progressBar = T
)
}
\arguments{
\item{raw_edges}{List of edges}

\item{file}{File from which edgelist can be read (specify only one file or raw_edges)}

\item{sep}{Field separator character used in read.table. If missing is set at " ".}

\item{header}{A logical value indicating whether "file" contains the names of the variables as its first line. If missing is set at F}

\item{outputFolder}{Folder to output results i.e. csv's, png's and the pdf.}

\item{csv}{A logical value indicating whether building blocks need to be outputted as csv files.}

\item{png}{A logical value indicating whether building blocks need to be outputted as png files.}

\item{pdf}{A logical value indicating whether building blocks need to be put together in a pdf file.}

\item{progressBar}{A logical value indicating whether to show a progress bar.}
}
\value{
A list of building blocks
}
\description{
This function finds fibration building blocks Specify only one file or raw_edges.
}
