% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coloring.R
\name{get.input.set.color.vector}
\alias{get.input.set.color.vector}
\title{Find the input set color vector given the coloring}
\usage{
get.input.set.color.vector(
  nodes,
  raw_edges = NA,
  file = NA,
  sep = "\\t",
  header = F,
  directed = F
)
}
\arguments{
\item{nodes}{dataframe with 2 columns: Name, Color.}

\item{raw_edges}{List of edges}

\item{file}{File from which edgelist can be read (specify only one file or raw_edges)}

\item{sep}{Field separator character used in read.table. If missing is set at "\\t".}

\item{header}{A logical value indicating whether "file" contains the names of the variables as its first line. If missing is set at F}

\item{directed}{A logical value indicating whether the network is directed or not. If missing is set at F}
}
\value{

}
\description{
This function finds the input set color vector given the coloring. Note, can be undirected
}
